requirements.txt
aiofiles
aiosmtplib
annotated-types
anyio
asyncio
bcrypt
blinker
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
cssselect
cssutils
dnspython
ecdsa
email_validator
fastapi
fastapi-mail
firebase-admin
gcloud
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httplib2
idna
Jinja2
jose
jwcrypto
lxml
Mako
MarkupSafe
msgpack
Naked
oauth2client
passlib
pip
premailer
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pycryptodome
pydantic
pydantic_core
pydantic-settings
PyJWT
pyotp
pyparsing
Pyrebase4
python-dotenv
python-jose
python-jwt
python-multipart
PyYAML
requests
requests-toolbelt
rsa
setuptools
shellescape
six
sniffio
SQLAlchemy
starlette
typing_extensions
uritemplate
urllib3
uvicorn
yagmail
